script minor fix

+++ 42_docker/requirements/wordpress/wp_setup.sh
@@ -2,14 +2,17 @@
 
 echo "wordpress script start"
 
+if [ -f ./wp-activate.php ]
+then
+	echo "wordpress already downloaded"
+else
 
-
-wget http://wordpress.org/latest.tar.gz
-tar xfz latest.tar.gz
-mv wordpress/* .
-rm -rf latest.tar.gz
-rm -rf wordpress
-
+	wget http://wordpress.org/latest.tar.gz
+	tar xfz latest.tar.gz
+	mv wordpress/* .
+	rm -rf latest.tar.gz
+	rm -rf wordpress
+fi
 
 echo "wordpress script end"
